+++ src/backend/sse2/f32.rs
@@ -7,7 +7,7 @@ use core::arch::x86_64::*;
 use crate::utils::PrimitiveFloatUtils;
 use crate::{
     Aligned, Mask, Mask3A, Mask4A, QuatA, Quaternion, Vec3A, Vec4A, Vector,
-    backend::{FloatVectorBackend, MaskBackend, QuaternionBackend, VectorBackend},
+    backend::{AffineBackend, FloatVectorBackend, MaskBackend, QuaternionBackend, VectorBackend},
     utils::safe_target_feature,
 };
 
@@ -266,6 +266,15 @@ unsafe impl VectorBackend<4, Aligned> for f32 {
             Mask(_mm_cmpge_ps(vector.0, other.0))
         }
     }
+}
+
+// SAFETY: The first `__m128` represents the matrix, the first two elements of
+// the second `__m128` represent the vector, and the two remaining elements are
+// padding. The padding satisfies the requirements of `Pod`. `Mat2A<f32>` is
+// represented by `Vec4A<f32>` which is represented by `__m128`, so we have the
+// same alignment.
+unsafe impl AffineBackend<2, Aligned> for f32 {
+    type Inner = [__m128; 2];
 }
 
 impl QuaternionBackend<Aligned> for f32 {