+++ src/scalar.rs
@@ -1,6 +1,6 @@
 use crate::{
     Aligned, Alignment, Unaligned,
-    backend::{DefaultBackend, MaskBackend, QuaternionBackend, VectorBackend},
+    backend::{AffineBackend, DefaultBackend, MaskBackend, QuaternionBackend, VectorBackend},
 };
 
 /// A trait for elements of vectors.
@@ -30,6 +30,12 @@ pub trait Scalar:
     + VectorBackend<2, Unaligned>
     + VectorBackend<3, Unaligned>
     + VectorBackend<4, Unaligned>
+    + AffineBackend<2, Aligned>
+    + AffineBackend<3, Aligned>
+    + AffineBackend<4, Aligned>
+    + AffineBackend<2, Unaligned>
+    + AffineBackend<3, Unaligned>
+    + AffineBackend<4, Unaligned>
     + QuaternionBackend<Aligned>
     + QuaternionBackend<Unaligned>
     + MaskBackend<2, Aligned>